+++ src/main/java/com/presentation/demo/repository/UserRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
-    User findUserById(Integer id);
+    User findUserById(Long id);
     User findUserByUsername(String username);
 }
